checkpoint: multiple inheritance init at least runs at this point
Lots of headache with multiple inheritance, specifically when
MMRAverager inherits from both MM_base and RAverager classes.
  - In the older implementation when MM_base didn't inherit from
QickProgram, the __init__ method of MM_base never gets called! Because
the method resolution order (MRO) of MMAverager goes like [MMAverager,
RAverager, QickProgram, AbsQickProgram, MM_base, object], but
AbsQickProgram never calls super().__init__() so MM_base.__init__()
never gets called.
  - I added QickProgram as MM_base's parent so that the type checker
knows where the methods such as us2cycles() are coming from. But this
alters the inheritance tree and the MRO becomes [MMRAverager, RAverager,
MM_base, QickProgram, AbsQickProgram, object]. This is problematic
because when RAverager calls super().__init()__, it intends to call its
parent QickProgram so only passes soccfg and not cfg, but this is
disrupted by the insertion of MM_base into the MRO list.
  - Two solutions: 1) keep the old inheritance tree and explicitly call
MM_base().__init() in the multiply inherited children; 2) alter the
inheritance tree. I started playing around with 2 but remember that
maybe 1 is eventually the better option?
  - Anyway, now attempting this approach: reverse the inheritance order
of MM_base and RAverager so that the super() inside RAverager still
finds its parent. Call the parent __init__ methods by name explicitly in
the children and patch all the little paper cuts manually (like copying
cfg.expt to the top level so that reps and rounds and expts are
available etc). At least the __init__ methods can finish running now...
  - Perhaps give approach 1) another try? It's less "good" code (MM_base
won't work if you don't multiply inherit it together with other
QickPrograms and type checkers complain), but maybe causes the least
disruption. As of this commit, the rb and dual rail classes are probably
broken because of the change to MM_base too.