aftermath of channel switching
Goodbye multimode 216 (Qick101): while cable (un)plugging, some metal
must have touched and shorted something and led to the failure of some
parts (see #fail on Slack) which eventually triggered a fire on board.
Replacing with the Qram 216 board for the moment, but this requires
getting the Qick repos on board and on the meas PC in sync. Simplest
solution is to abandon all the custom modifications on the PC and pull
the repo anew. This has the side effect of breaking old code that
depends on those custom modifications in the local qick repo. Most
commonly, the acquire function of the QickProgram only took the debug
kwarg in the modified qick asm file. All the downstream acquire
functions that pass this argument have to be either updated to remove
this argument or reference a new wrapped acquire function in MMAverager
or MMRAverager that accepts the debug argument. Also some imports that
no longer work are commented out