vendoring in code from slab
The status of the slab repo on the BF5 PC is beyond my ability to repair. 300 untracked changes the messiest part of which seemingly includes large scale, but partial, modifications of a proprietary Alazar SDK. After a full day wrestling with the history trying to straighten things out, I decided it's just not worth it. We mostly use RFSoC for this project now and the absolute majority of the slab repo is simply not unused. The tech debt was accumulated over many years and not worth the effort to salvage. So just copied in the files we actively use in this repo into the slab module HERE and made necessary changes to make stuff run. pyproject.toml no longer requires a separate slab package.